+++ .github/workflows/main.yml
@@ -27,10 +27,10 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install -U pip
-          pip install -r src/python/requirements.txt
+          pip install -r tools/python/requirements.txt
       - name: Synchronize the csv file with an up-to-date one
         run: |
-          cd src/python
+          cd tools/python
           python download.py
           python csv-json.py "../csv/kdb-$(date +%Y%m%d).csv" list.txt
           cd ../..
