--- src/python/list.txt
@@ -1,104 +1,0 @@
-1/2/3/4/5/6, 総合科目
-	11, フレッシュマン・セミナー
-	1227/1228, 学問への誘い
-	12[1227/1228], 学士基盤科目
-	14, 学士基盤科目（高年次向け）
-
-	2, 体育
-		21, 1年次
-		22, 2年次
-		23, 3年次
-
-	3, 外国語
-		31[3190012/3190022], 英語
-		32, ドイツ語
-		33, フランス語
-		34, 中国語
-		35, ロシア語
-		36, スペイン語
-		37, 朝鮮語
-		391/392, 日本語、Japan-Expert（学士）プログラム
-		390, 外国語（短期留学生用カリキュラム）
-
-	6, 情報
-		61, 情報リテラシー（講義）
-		64, 情報リテラシー（演習）
-		65, データサイエンス
-
-	5, 国語
-	4, 芸術
-
-8/9, 自由科目
-	8[8049911/8049921/8149911], 自由科目（特設）
-	9[99], 教職に関する科目
-	99, 博物館に関する科目
-	8049911/8049921/8149911, 日本事情等科目
-
-A/B/C/E/F/G/H/W/Y/VB, 専門基礎科目・専門科目
-	AB0/BB0, 全学群対象
-
-	A[AB0], 人文・文化学群
-		AB[AB0], 人文学類
-		AC, 比較文化学類	
-		AE, 日本語・日本文化学類
-
-	B[BE], 社会・国際学群
-		BA, 社会・国際学群グローバル科目群
-		BB[BB00101/BB00201/BB00301/BB00401/BB00501/BB00601], 社会学類
-		BC, 国際総合学類
-
-	C, 人間学群
-		CA, 学群コア・カリキュラム
-		CB, 教育学類
-		CC, 心理学類
-		CE, 障害科学類
-
-	E, 生命環境学群
-		EZ, 学群共通科目
-		EB, 生物学類
-		EC, 生物資源学類
-		EE, 地球学類
-
-	F, 理工学群
-		FA, 学群共通科目
-		FB, 数学類
-		FC, 物理学類
-		FE, 化学類
-		FF/FA01131/FA01141/FA01231/FA01241/FA01331/FA01341/FA01431/FA01441/FA01531/FA01541/FA01631/FA01641/FA01731/FA01741/FA01831/FA01841, 応用理工学類
-		FG, 工学システム学類
-		FH, 社会工学類
-
-	G, 情報学群
-		GA, 学群共通科目
-		GB, 情報科学類
-		GC, 情報メディア創成学類
-		GE, 知識情報・図書館学類
-
-	H, 医学群
-		HB/FCB1231/FCB1331/FE11181/FE11191, 医学類
-		HC, 看護学類
-		HE, 医療科学類
-
-	W, 体育専門学群
-
-	Y, 芸術専門学群
-		YA, 専門基礎科目(共通)
-		Y11/YBA, 専門科目(美術史領域)
-		Y12/YBB, 専門科目(芸術支援領域)
-		YBC, 専門科目(洋画領域)
-		YBD, 専門科目(版画領域)
-		YBE, 専門科目(日本画領域)
-		Y22/YBF, 専門科目(彫塑領域)
-		YBG, 専門科目(書領域)
-		Y63/YBH, 専門科目(クラフト領域)
-		YBJ, 専門科目(総合造形領域)
-		YBK, 専門科目(構成領域)
-		YBL, 専門科目(ビジュアルデザイン領域)
-		YBN, 専門科目(情報・プロダクトデザイン領域)
-		YBP, 専門科目(環境デザイン領域)
-		Y71/YBQ, 専門科目(建築デザイン領域)
-		Y61/Y62/YBW/YBX, 専門科目(共通)
-	
-	VB, グローバル教育院 地球規模課題学位プログラム
-
-XJ, 外国人留学生等日本語研修コース
